+++ packages/api-clients/api-codegen-preset/src/api-project.ts
@@ -1,8 +1,9 @@
 import fs from 'fs';
 
 import {pluckConfig} from '@shopify/graphql-codegen';
+import type {IGraphQLProject, SchemaPointer} from 'graphql-config';
 
-import type {ShopifyApiProjectOptions} from './types';
+import {type ShopifyApiProjectOptions} from './types';
 import {shopifyApiTypes} from './api-types';
 import {getSchemaData} from './helpers/get-schema-data';
 
@@ -13,13 +14,17 @@ export const shopifyApiProject = ({
   outputDir = '.',
   documents = ['**/*.{ts,tsx}', '!node_modules'],
   declarations = true,
-}: ShopifyApiProjectOptions) => {
-  const {schema, schemaFile} = getSchemaData(outputDir, apiType, apiVersion);
+  apiKey,
+}: ShopifyApiProjectOptions): IGraphQLProject => {
+  const {schema, schemaFile} = getSchemaData(outputDir, apiType, {
+    apiVersion,
+    apiKey,
+  });
 
   const schemaFileExists = fs.existsSync(`${schemaFile}`);
 
   return {
-    schema: schemaFileExists ? schemaFile : schema,
+    schema: schemaFileExists ? schemaFile : (schema as SchemaPointer),
     documents,
     extensions: {
       codegen: {
@@ -27,6 +32,7 @@ export const shopifyApiProject = ({
         generates: shopifyApiTypes({
           apiType,
           apiVersion,
+          apiKey,
           outputDir,
           documents,
           module,