+++ packages/api-clients/api-codegen-preset/src/api-types.ts
@@ -1,7 +1,9 @@
 import fs from 'fs';
 
+import type {CodegenConfig} from '@graphql-codegen/cli';
+
 import {preset} from './preset';
-import type {ShopifyApiTypesOptions} from './types';
+import {type ShopifyApiTypesOptions} from './types';
 import {getSchemaData} from './helpers/get-schema-data';
 import {getOutputFiles} from './helpers/get-output-files';
 
@@ -12,8 +14,12 @@ export const shopifyApiTypes = ({
   outputDir = '.',
   documents = ['**/*.{ts,tsx}', '!**/node_modules'],
   declarations = true,
-}: ShopifyApiTypesOptions) => {
-  const {schema, schemaFile} = getSchemaData(outputDir, apiType, apiVersion);
+  apiKey,
+}: ShopifyApiTypesOptions): CodegenConfig['generates'] => {
+  const {schema, schemaFile} = getSchemaData(outputDir, apiType, {
+    apiVersion,
+    apiKey,
+  });
   const {typesFile, queryTypesFile} = getOutputFiles(apiType, declarations);
 
   const schemaFileExists = fs.existsSync(`${schemaFile}`);