+++ packages/apps/shopify-api/lib/billing/index.ts
@@ -5,6 +5,7 @@ import {check} from './check';
 import {request} from './request';
 import {cancel} from './cancel';
 import {subscriptions} from './subscriptions';
+import {createUsageRecord} from './create-usage-record';
 import {ShopifyBilling} from './types';
 
 export type {ShopifyBilling} from './types';
@@ -17,5 +18,6 @@ export function shopifyBilling<Future extends FutureFlagOptions>(
     request: request(config),
     cancel: cancel(config),
     subscriptions: subscriptions(config),
+    createUsageRecord: createUsageRecord(config),
   };
 }