+++ packages/apps/shopify-app-remix/src/server/__test-helpers/get-jwt.ts
@@ -1,31 +1,10 @@
 import {JwtPayload} from '@shopify/shopify-api';
-import jwt from 'jsonwebtoken';
+import {getJwt as getJwtImport} from '@shopify/shopify-api/test-helpers';
 
-import {API_KEY, API_SECRET_KEY, TEST_SHOP, USER_ID} from './const';
+import {API_KEY, API_SECRET_KEY, TEST_SHOP_NAME} from './const';
 
-interface TestJwt {
-  token: string;
-  payload: JwtPayload;
-}
-
-export function getJwt(overrides: Partial<JwtPayload> = {}): TestJwt {
-  const date = new Date();
-  const payload = {
-    iss: `${TEST_SHOP}/admin`,
-    dest: `https://${TEST_SHOP}`,
-    aud: API_KEY,
-    sub: `${USER_ID}`,
-    exp: date.getTime() / 1000 + 3600,
-    nbf: date.getTime() / 1000 - 3600,
-    iat: date.getTime() / 1000 - 3600,
-    jti: '1234567890',
-    sid: '0987654321',
-    ...overrides,
-  };
-
-  const token = jwt.sign(payload, API_SECRET_KEY, {
-    algorithm: 'HS256',
-  });
-
-  return {token, payload};
+export function getJwt(
+  overrides: Partial<JwtPayload> = {},
+): ReturnType<typeof getJwtImport> {
+  return getJwtImport(TEST_SHOP_NAME, API_KEY, API_SECRET_KEY, overrides);
 }