+++ packages/apps/shopify-app-remix/src/server/authenticate/admin/billing/__tests__/mock-responses.ts
@@ -1,3 +1,5 @@
+import {Money} from '@shopify/shopify-api';
+
 export const PLAN_1 = 'Shopify app plan 1';
 export const PLAN_2 = 'Shopify app plan 2';
 export const ALL_PLANS = [PLAN_1, PLAN_2];
@@ -78,3 +80,93 @@ export const CANCEL_RESPONSE = JSON.stringify({
     },
   },
 });
+
+export const SUBSCRIPTIONS_WITH_USAGE_PLANS_RESPONSE = JSON.stringify({
+  data: {
+    currentAppInstallation: {
+      oneTimePurchases: {
+        edges: [],
+        pageInfo: {
+          hasNextPage: false,
+          endCursor: null,
+        },
+      },
+      activeSubscriptions: [
+        {
+          id: 'gid://123',
+          test: true,
+          status: 'ACTIVE',
+          name: 'Basic Usage Plan',
+          lineItems: [
+            {
+              id: 'gid://456',
+              plan: {
+                pricingDetails: {
+                  balanceUsed: {
+                    amount: 0,
+                    currencyCode: 'USD',
+                  },
+                  cappedAmount: {
+                    amount: '5.00',
+                    currencyCode: 'USD',
+                  },
+                  terms: '1 dollar per usage',
+                },
+              },
+            },
+          ],
+        },
+      ],
+      hasActivePayment: true,
+    },
+  },
+});
+
+export const USAGE_RECORD_DESCRIPTION = 'Usage record description';
+export const USAGE_RECORD_PRICE = {
+  amount: 1.0,
+  currencyCode: 'USD',
+} as Money;
+export const USAGE_RECORD_SUBSCRIPTION_ID = 'gid://456';
+
+export const USAGE_RECORD_CREATE_RESPONSE = JSON.stringify({
+  data: {
+    appUsageRecordCreate: {
+      appUsageRecord: {
+        id: 'gid://123',
+        description: USAGE_RECORD_DESCRIPTION,
+        price: USAGE_RECORD_PRICE,
+        subscriptionLineItem: {
+          id: USAGE_RECORD_SUBSCRIPTION_ID,
+          plan: {
+            pricingDetails: {
+              balanceUsed: USAGE_RECORD_PRICE,
+              cappedAmount: {
+                amount: '5.00',
+                currencyCode: 'USD',
+              },
+              terms: '1 dollar per usage',
+            },
+          },
+        },
+      },
+      userErrors: [],
+    },
+  },
+});
+
+export const USAGE_RECORD = {
+  description: 'Usage record description',
+  id: 'gid://123',
+  price: {amount: 1, currencyCode: 'USD'},
+  subscriptionLineItem: {
+    id: 'gid://456',
+    plan: {
+      pricingDetails: {
+        balanceUsed: {amount: 1, currencyCode: 'USD'},
+        cappedAmount: {amount: '5.00', currencyCode: 'USD'},
+        terms: '1 dollar per usage',
+      },
+    },
+  },
+};