+++ packages/apps/shopify-app-remix/src/server/authenticate/admin/billing/index.ts
@@ -2,3 +2,4 @@ export {cancelBillingFactory} from './cancel';
 export {requireBillingFactory} from './require';
 export {requestBillingFactory} from './request';
 export {checkBillingFactory} from './check';
+export {createUsageRecordFactory} from './create-usage-record';