+++ packages/apps/shopify-app-remix/src/server/authenticate/admin/scope/__tests__/query.test.ts
@@ -25,9 +25,9 @@ it('returns scopes information', async () => {
 
   // THEN
   expect(result).not.toBeUndefined();
-  expect(result.granted.required).toEqual(['read_orders']);
-  expect(result.granted.optional).toEqual(['write_customers']);
-  expect(result.declared.required).toEqual(['write_orders', 'read_reports']);
+  expect(result.granted).toEqual(['read_orders', 'write_customers']);
+  expect(result.required).toEqual(['read_orders', 'read_reports']);
+  expect(result.optional).toEqual(['write_customers']);
 });
 
 it('redirects to exit-iframe with authentication using app bridge when embedded and Shopify invalidated the session', async () => {